Fix overflow issues on safari
